# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "batsim",
    "description": "Residential battery fleet simulator: physics-faithful virtual homes behind an OpenAPI-first HTTP API for dispatch-strategy testing (ERCOT).",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/dispatch": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Send a command to a home/fleet subset.",
        "operationId": "dispatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Command accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown target",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Empty target set",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/homes": {
      "get": {
        "tags": [
          "homes"
        ],
        "summary": "List homes.",
        "operationId": "list_homes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continuation cursor.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(?:[A-Za-z0-9_-]{4})*(?:[A-Za-z0-9_-]{2,3})?$"
            }
          },
          {
            "name": "fleet_id",
            "in": "query",
            "description": "Restrict to a fleet.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "Restrict to an operating mode.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OperatingMode"
            }
          },
          {
            "name": "load_zone",
            "in": "query",
            "description": "Restrict to a load zone.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "battery_model",
            "in": "query",
            "description": "Restrict to a battery model.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of homes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomesPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "homes"
        ],
        "summary": "Create a home.",
        "operationId": "create_home",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Home created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeDoc"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Fleet not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Composition rule violation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/homes/{id}": {
      "get": {
        "tags": [
          "homes"
        ],
        "summary": "Get one home.",
        "operationId": "get_home",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Home id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The home",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown home",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "homes"
        ],
        "summary": "Delete a home.",
        "operationId": "delete_home",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Home id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Home deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown home",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Simulation is running",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "homes"
        ],
        "summary": "Update mutable home configuration.",
        "operationId": "patch_home",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Home id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchHomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated home",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeDoc"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown home",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Simulation is running",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionSpec": {
        "oneOf": [
          {
            "type": "object",
            "description": "Charge at a fixed power.",
            "required": [
              "kw",
              "type"
            ],
            "properties": {
              "duration_s": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "Hold duration in seconds; the home returns to\nself-consumption afterwards. Omit to hold indefinitely.",
                "maximum": 18446744073709551615,
                "minimum": 0
              },
              "kw": {
                "type": "number",
                "format": "double",
                "description": "Charge power in kW (> 0).",
                "exclusiveMinimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "charge_to"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Discharge at a fixed power.",
            "required": [
              "kw",
              "type"
            ],
            "properties": {
              "duration_s": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "Hold duration in seconds; the home returns to\nself-consumption afterwards. Omit to hold indefinitely.",
                "maximum": 18446744073709551615,
                "minimum": 0
              },
              "kw": {
                "type": "number",
                "format": "double",
                "description": "Discharge power in kW (> 0).",
                "exclusiveMinimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "discharge_to"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Set the backup reserve floor.",
            "required": [
              "soc",
              "type"
            ],
            "properties": {
              "soc": {
                "type": "number",
                "format": "double",
                "description": "Reserve as a fraction of usable energy (0..1).",
                "maximum": 1,
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "set_reserve_soc"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Switch operating mode.",
            "required": [
              "mode",
              "type"
            ],
            "properties": {
              "mode": {
                "$ref": "#/components/schemas/OperatingMode",
                "description": "The new mode."
              },
              "type": {
                "type": "string",
                "enum": [
                  "set_mode"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Curtail PV output.",
            "required": [
              "pct",
              "type"
            ],
            "properties": {
              "pct": {
                "type": "number",
                "format": "double",
                "description": "Curtailed fraction of output in percent (0..100).",
                "maximum": 100,
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "curtail_pv"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Clear all overrides: self-consumption mode, no setpoint, no\ncurtailment.",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "clear_override"
                ]
              }
            }
          }
        ],
        "description": "The dispatch action. Closed union, discriminated by `type`."
      },
...
      "CommandStatus": {
        "type": "string",
        "description": "Command rollup status.",
        "enum": [
          "queued",
          "in_flight",
          "completed",
          "completed_with_errors",
          "cancelled"
        ]
      },
...
      "CreateHomeRequest": {
        "type": "object",
        "description": "Create-home request.",
        "required": [
          "battery",
          "load",
          "location"
        ],
        "properties": {
          "battery": {
            "$ref": "#/components/schemas/BatterySpec",
            "description": "Battery selection."
          },
          "fleet_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Fleet to attach the home to (must exist when given)."
          },
          "initial_soc": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Initial state of charge (fraction of usable; default 0.5).",
            "maximum": 1,
            "minimum": 0
          },
          "inverter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InverterSpec",
                "description": "Inverter selection (optional; vendor default composed when\nomitted)."
              }
            ]
          },
          "load": {
            "$ref": "#/components/schemas/LoadSpec",
            "description": "Household load."
          },
          "location": {
            "$ref": "#/components/schemas/LocationSpec",
            "description": "Siting."
          },
          "pv": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PvSpec",
                "description": "PV array (optional)."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "DispatchRequest": {
        "type": "object",
        "description": "Dispatch request.",
        "required": [
          "target",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionSpec",
            "description": "Action."
          },
          "command_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-supplied command id (ULID/uuid). Retries with the same id\nare deduplicated before enqueue. Server-assigned when omitted."
          },
          "execution": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExecutionSpec",
                "description": "Execution shaping."
              }
            ]
          },
          "target": {
            "$ref": "#/components/schemas/TargetSpec",
            "description": "Target set."
          }
        },
        "additionalProperties": false
      },
      "DispatchResponse": {
        "type": "object",
        "description": "Dispatch acceptance response.",
        "required": [
          "command_id",
          "accepted",
          "targets",
          "status",
          "status_url"
        ],
        "properties": {
          "accepted": {
            "type": "boolean",
            "description": "Always true for a 202 response."
          },
          "command_id": {
            "type": "string",
            "description": "The command id (echoed or server-assigned)."
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus",
            "description": "Rollup status at response time."
          },
          "status_url": {
            "type": "string",
            "description": "Poll URL for execution detail."
          },
          "targets": {
            "type": "integer",
            "description": "Number of homes targeted.",
            "minimum": 0
          }
        }
      },
...
      "HomeDoc": {
        "type": "object",
        "description": "Home document.",
        "required": [
          "id",
          "config",
          "state",
          "created_at"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/HomeConfigDoc",
            "description": "Validated configuration."
          },
          "created_at": {
            "type": "string",
            "description": "Wall-clock creation time."
          },
          "id": {
            "type": "string",
            "description": "Home id."
          },
          "state": {
            "$ref": "#/components/schemas/HomeStateDoc",
            "description": "Live state."
          }
        }
      },
...
      "HomesPage": {
        "type": "object",
        "description": "Page of homes.",
        "required": [
          "data",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HomeDoc"
            },
            "description": "Homes."
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination."
          }
        }
      },
...
      "OkDoc": {
        "type": "object",
        "description": "Empty success body for action endpoints with nothing to return.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Always true."
          }
        }
      },
      "OperatingMode": {
        "type": "string",
        "description": "Operating modes a home can be commanded into.",
        "enum": [
          "self-consumption",
          "backup-only",
          "time-of-use",
          "grid-services"
        ]
      },
...
      "PatchHomeRequest": {
        "type": "object",
        "description": "Mutable home configuration.",
        "properties": {
          "mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OperatingMode",
                "description": "New operating mode."
              }
            ]
          },
          "reserve_soc": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "New backup reserve floor (fraction of usable).",
            "maximum": 1,
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem document; the only error body the API emits.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Specifics of this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request correlation id (echoed as `X-Request-Id`)."
          },
          "type": {
            "type": "string",
            "description": "Problem type URN."
          }
        },
        "additionalProperties": true
      },
...
    }
  },
...
}